Detect_API/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
ultralytics==8.0.196
Pillow==10.0.0
python-multipart